requirements.txt
Django>=5.1,<6.0
python-dotenv>=1.0.1
psycopg2-binary>=2.9.9
torch>=2.11.0
torchvision>=0.26.0
pillow>=12.2.0
ultralytics>=8.4.38
opencv-python>=4.13.0.92
requests>=2.33.0
pymysql>=1.1.1
gunicorn>=23.0.0
whitenoise>=6.6.0
